My c commit

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ long int li;
 long long int ll;
 float f;
 
-printf("Size of a char: %lu byte(s)\n", sizeof(c));
-printf("Size of an int :%lu bytes(s)\n", sizeof(i));
-printf("Size of long int: %lu byte(s)\n", sizeof(li));
-printf("Size of a long long int: %lu byte(s)\n", sizeof(ll));
-printf("Size of a float: %lu byte(s)\n", sizeof(f));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+printf("Size of an int :%lu bytes(s)\n", (unsigned long)sizeof(i));
+printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(li));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(ll));
+printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 
 return (0);
 }
